+++ lib/ioctl_wrappers.h
@@ -60,11 +60,11 @@ uint32_t __gem_create(int fd, int size);
 uint32_t gem_create(int fd, uint64_t size);
 void gem_execbuf(int fd, struct drm_i915_gem_execbuffer2 *execbuf);
 
-void *gem_mmap__gtt(int fd, uint32_t handle, int size, int prot);
-void *gem_mmap__cpu(int fd, uint32_t handle, int offset, int size, int prot);
+void *gem_mmap__gtt(int fd, uint32_t handle, uint64_t size, unsigned prot);
+void *gem_mmap__cpu(int fd, uint32_t handle, uint64_t offset, uint64_t size, unsigned prot);
 
 bool gem_mmap__has_wc(int fd);
-void *gem_mmap__wc(int fd, uint32_t handle, int offset, int size, int prot);
+void *gem_mmap__wc(int fd, uint32_t handle, uint64_t offset, uint64_t size, unsigned prot);
 
 /**
  * gem_require_mmap_wc:
